requirements.txt
Fabric
Flask-SQLAlchemy
Flask-WTF
WTForms
coverage
# ecdsa
# paramiko
# pycrypto
email_validator
flask-sqlacodegen
psycopg2
Flask-Login == 0.5.0
